+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Tabs can be rearranged
 - Settings for new tab logo and welcome message
 - Devtools now work
+- Tab name updates correctly for internal pages
 
 ## v0.7.0 - 2026-02-15